docker: refactor the binding with the local user
Using the --user option in the docker run allows to run a command using
the specified uid / gid. But the specific uid / gid does not map to any
users / groups declared in the passwd / group files. So if something
want to verify the user (like sudo) the /etc files needs to be mounted.

This is not very clean, and this is not working if you are using a non
local user (LDAP).

Using an entrypoint which create a new user / group with the specified
uid / gid / groups fix the issue.