oe-init-build-env: introduce the BB_EXPORT_LIST_VARIABLE variable
Some bitbake / yocto variables has been renamed to follow the inclusive
naming initiative. The variable BB_ENV_EXTRAWHITE is one of them and
has been renamed to BB_ENV_PASSTHROUGH_ADDITIONS.

Now, the variable to be used must be configurable to be able to build
a project prior to this change or not.