common: improve config parsing
When using "${MAKE} -C <subdir>" commands in the .config file it seems
that the entire sub-makefile is also parsed. This can lead to parsing
failures.

To avoid this, the ${MAKE} variable is forced to "true", so make is
happy when it prints its database, and the extracted commands still
have the ${MAKE} variable as is.