docker: refactor signal handling
In order to be able to pass the signals to the final command correctly,
the "exec" form must be used for each step, instead of the
"shell -c" form.

This causes "interrupted" error messages to be displayed again as we
cannot mute them anymore, but this is a lesser evil.